RF: refactored sqrt_lasso and moved old module to sandbox
